✨ 🚧 Work on parsing class for all types of text
Parser will handle all types of parsing text.
The Parser is used so that we don't have to implement a new class
for every different type of document we want to parse.
Instead we can simply add a new parser object in the DB and
that object can check if it should parse the text and then do that.